move out rec logic to own class
